Extract src/inspector.ts — and tick the render-modules item
The module the item named but neither evidence group listed, because
./inspectorview already held every card's markup. What was left in main.ts is
the part that isn't a pure function: the dispatcher, the status pill, and the
two cards that are *not* an agent's — the shell card, and the task-run card
with its four per-Sess button listeners (which address one specific run and
so were never part of the global [data-*] dispatcher).

renderInspector moved to the top of the new module as the dispatcher it is;
that reorder is the only structural change, and the verification compares the
two blocks line-sorted to prove it: identical modulo one `export` and three
`host.` prefixes.

Three callees — rerunTask, revealSource, sendOutputToSession — all belong to
the run machinery that owns panes and backend calls. At the ~4 boundary PLAN
draws, so one host rather than three setters.

**With this the render-modules item is done, and PLAN says what stayed behind
and why** rather than leaving it to be rediscovered: renderHeader and
syncStageButtons (10 lines each — a stage.ts would be a file per twenty
lines); the read-only external/dormant mirrors, which are not a render cluster
but welded to the openExternal/openDormant/loadTranscript machinery beside
them; and renderAll, which is the next item's subject.

main.ts 5,705 -> 1,593 lines (-72%) across the whole phase; 1,682 -> 1,593
here. Gates: tsc --noEmit clean, 368 vitest pass. No Rust touched.
